# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flashcard app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Card {
  id: number;
  front: string;
  back: string;
}

export interface CardListInput {
  cards: Card[];
}

export type ResultsInputResults = {[key: string]: boolean};

export interface ResultsInput {
  results: ResultsInputResults;
}

export interface SaveResult {
  status: string;
  /** @nullable */
  message?: string | null;
}