# lib/api-zod/src/generated/types/cardListInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flashcard app API
 * OpenAPI spec version: 0.1.0
 */
import type { Card } from './card';

export interface CardListInput {
  cards: Card[];
}