# lib/api-zod/src/generated/types/resultsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flashcard app API
 * OpenAPI spec version: 0.1.0
 */
import type { ResultsInputResults } from './resultsInputResults';

export interface ResultsInput {
  results: ResultsInputResults;
}