# lib/api-zod/src/generated/types/saveResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flashcard app API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveResult {
  status: string;
  /** @nullable */
  message?: string | null;
}